ScreenModel/FastAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
joblib
pandas
numpy
scikit-learn==1.6.1
xgboost
dotenv
google-genai
